+++ src/main/scala/edu/ie3/simona/agent/participant2/ParticipantAgentInit.scala
@@ -25,7 +25,7 @@ import edu.ie3.simona.ontology.messages.services.WeatherMessage.RegisterForWeath
 import edu.ie3.simona.ontology.messages.{Activation, SchedulerMessage}
 import edu.ie3.simona.service.ServiceType
 import edu.ie3.simona.util.SimonaConstants.INIT_SIM_TICK
-import org.apache.pekko.actor.typed.scaladsl.Behaviors
+import org.apache.pekko.actor.typed.scaladsl.{ActorContext, Behaviors}
 import org.apache.pekko.actor.typed.scaladsl.adapter.TypedActorRefOps
 import org.apache.pekko.actor.typed.{ActorRef, Behavior}
 import org.apache.pekko.actor.{ActorRef => ClassicRef}
@@ -216,7 +216,7 @@ object ParticipantAgentInit {
         parentData,
       )
 
-    case (_, RegistrationFailedMessage(_)) =>
+    case (ctx, RegistrationFailedMessage(_)) =>
       // we're _not_ supposed to replay primary data, thus initialize the physical model
       val modelShell = ParticipantModelShell.createForPhysicalModel(
         participantInput,
@@ -256,7 +256,7 @@ object ParticipantAgentInit {
             modelShell,
             serviceType,
             serviceRef,
-          )
+          )(ctx)
         }
 
         waitingForServices(
@@ -275,7 +275,7 @@ object ParticipantAgentInit {
       modelShell: ParticipantModelShell[_, _],
       serviceType: ServiceType,
       serviceRef: ClassicRef,
-  ): Unit =
+  )(implicit ctx: ActorContext[ParticipantAgent.Request]): Unit =
     serviceType match {
       case ServiceType.WeatherService =>
         val geoPosition = participantInput.getNode.getGeoPosition
@@ -298,7 +298,7 @@ object ParticipantAgentInit {
         )
 
       case ServiceType.EvMovementService =>
-        serviceRef ! RegisterForEvDataMessage(modelShell.uuid)
+        serviceRef ! RegisterForEvDataMessage(ctx.self, modelShell.uuid)
     }
 
   /** Waiting for replies from secondary services. If all replies have been