+++ src/main/scala/edu/ie3/simona/ontology/messages/services/ServiceMessage.scala
@@ -6,18 +6,43 @@
 
 package edu.ie3.simona.ontology.messages.services
 
-import org.apache.pekko.actor.ActorRef
+import org.apache.pekko.actor.{ActorRef => ClassicRef}
 
 import java.util.UUID
-import edu.ie3.simona.agent.participant.data.Data
+import edu.ie3.simona.api.data.ontology.DataMessageFromExt
+import edu.ie3.simona.ontology.messages.Activation
+import edu.ie3.simona.ontology.messages.services.EvMessage.EvInternal
 import edu.ie3.simona.scheduler.ScheduleLock.ScheduleKey
+import edu.ie3.simona.service.ServiceStateData.InitializeServiceStateData
 
 /** Collections of all messages, that are send to and from the different
   * services
   */
-sealed trait ServiceMessage
+sealed trait ServiceMessage extends EvInternal
 
 object ServiceMessage {
+
+  final case class WrappedActivation(activation: Activation)
+      extends ServiceMessage
+
+  final case class WrappedExternalMessage(
+      extMsg: DataMessageFromExt
+  ) extends ServiceMessage
+
+  /** Service initialization data can sometimes only be constructed once the
+    * service actor is created (e.g.
+    * [[edu.ie3.simona.service.ev.ExtEvDataService]]). Thus, we need an extra
+    * initialization message.
+    */
+  final case class Create[+I <: InitializeServiceStateData](
+      initializeStateData: I,
+      unlockKey: ScheduleKey,
+  ) extends ServiceMessage
+
+  final case class ScheduleServiceActivation(
+      tick: Long,
+      unlockKey: ScheduleKey,
+  ) extends ServiceMessage
 
   /** Message used to register for a service
     */
@@ -29,7 +54,7 @@ object ServiceMessage {
     *   Identifier of the input model
     */
   final case class PrimaryServiceRegistrationMessage(
-      requestingActor: ActorRef,
+      requestingActor: ClassicRef,
       inputModelUuid: UUID,
   ) extends ServiceRegistrationMessage
 
@@ -40,12 +65,6 @@ object ServiceMessage {
     * @param requestingActor
     *   Reference to the requesting actor
     */
-  final case class WorkerRegistrationMessage(requestingActor: ActorRef)
+  final case class WorkerRegistrationMessage(requestingActor: ClassicRef)
       extends ServiceRegistrationMessage
-
-  final case class ScheduleServiceActivation(
-      tick: Long,
-      unlockKey: ScheduleKey,
-  )
-
 }