+++ src/main/scala/edu/ie3/simona/service/ev/ExtEvDataService.scala
@@ -6,6 +6,7 @@
 
 package edu.ie3.simona.service.ev
 
+import edu.ie3.simona.agent.participant2.ParticipantAgent
 import edu.ie3.simona.agent.participant2.ParticipantAgent.{
   DataProvision,
   RegistrationSuccessfulMessage,
@@ -21,36 +22,42 @@ import edu.ie3.simona.exceptions.{
   ServiceException,
 }
 import edu.ie3.simona.model.participant.evcs.EvModelWrapper
+import edu.ie3.simona.ontology.messages.services.EvMessage
 import edu.ie3.simona.ontology.messages.services.EvMessage._
-import edu.ie3.simona.ontology.messages.services.ServiceMessage.ServiceRegistrationMessage
+import edu.ie3.simona.ontology.messages.services.ServiceMessage.{
+  ServiceRegistrationMessage,
+  WrappedExternalMessage,
+}
 import edu.ie3.simona.service.ServiceStateData.{
   InitializeServiceStateData,
   ServiceBaseStateData,
 }
-import edu.ie3.simona.service.ev.ExtEvDataService.{
-  ExtEvStateData,
-  InitExtEvData,
+import edu.ie3.simona.service.{
+  ExtDataSupport,
+  ServiceStateData,
+  TypedSimonaService,
 }
-import edu.ie3.simona.service.{ExtDataSupport, ServiceStateData, SimonaService}
 import edu.ie3.simona.util.ReceiveDataMap
 import edu.ie3.simona.util.SimonaConstants.INIT_SIM_TICK
-import org.apache.pekko.actor.{ActorContext, ActorRef, Props}
+import org.apache.pekko.actor.typed.scaladsl.adapter.TypedActorRefOps
+import org.apache.pekko.actor.typed.scaladsl.{ActorContext, Behaviors}
+import org.apache.pekko.actor.typed.{ActorRef, Behavior}
+import org.slf4j.Logger
 
 import java.util.UUID
 import scala.jdk.CollectionConverters._
 import scala.jdk.OptionConverters._
 import scala.util.{Failure, Success, Try}
 
-object ExtEvDataService {
+object ExtEvDataService
+    extends TypedSimonaService[EvMessage]
+    with ExtDataSupport[EvMessage] {
 
-  def props(scheduler: ActorRef): Props =
-    Props(
-      new ExtEvDataService(scheduler: ActorRef)
-    )
+  override type S = ExtEvStateData
 
   final case class ExtEvStateData(
       extEvData: ExtEvDataConnection,
-      uuidToActorRef: Map[UUID, ActorRef] = Map.empty[UUID, ActorRef],
+      uuidToActorRef: Map[UUID, ActorRef[ParticipantAgent.Request]] = Map.empty,
       extEvMessage: Option[EvDataMessageFromExt] = None,
       freeLots: ReceiveDataMap[UUID, Int] = ReceiveDataMap.empty,
       departingEvResponses: ReceiveDataMap[UUID, Seq[EvModelWrapper]] =
@@ -61,12 +68,26 @@ object ExtEvDataService {
       extEvData: ExtEvDataConnection
   ) extends InitializeServiceStateData
 
-}
+  def adapter(evService: ActorRef[EvMessage]): Behavior[DataMessageFromExt] =
+    Behaviors.receiveMessagePartial { extMsg =>
+      evService ! WrappedExternalMessage(extMsg)
+      Behaviors.same
+    }
 
-class ExtEvDataService(override val scheduler: ActorRef)
-    extends SimonaService[ExtEvStateData](scheduler)
-    with ExtDataSupport[ExtEvStateData] {
-
+  /** Initialize the concrete service implementation using the provided
+    * initialization data. This method should perform all heavyweight tasks
+    * before the actor becomes ready. The return values are a) the state data of
+    * the initialized service and b) optional triggers that should be sent to
+    * the [[edu.ie3.simona.scheduler.Scheduler]] together with the completion
+    * message that is sent in response to the trigger that is sent to start the
+    * initialization process
+    *
+    * @param initServiceData
+    *   the data that should be used for initialization
+    * @return
+    *   the state data of this service and optional tick that should be included
+    *   in the completion message
+    */
   override def init(
       initServiceData: ServiceStateData.InitializeServiceStateData
   ): Try[
@@ -107,11 +128,12 @@ class ExtEvDataService(override val scheduler: ActorRef)
   override def handleRegistrationRequest(
       registrationMessage: ServiceRegistrationMessage
   )(implicit
-      serviceStateData: ExtEvStateData
-  ): Try[ExtEvStateData] =
+      serviceStateData: S,
+      ctx: ActorContext[EvMessage],
+  ): Try[S] =
     registrationMessage match {
-      case RegisterForEvDataMessage(evcs) =>
-        Success(handleRegistrationRequest(sender(), evcs))
+      case RegisterForEvDataMessage(actorRef, evcs) =>
+        Success(handleRegistrationRequest(actorRef, evcs))
       case invalidMessage =>
         Failure(
           InvalidRegistrationRequestException(
@@ -135,12 +157,13 @@ class ExtEvDataService(override val scheduler: ActorRef)
     *   information if the registration has been carried out successfully
     */
   private def handleRegistrationRequest(
-      agentToBeRegistered: ActorRef,
+      agentToBeRegistered: ActorRef[ParticipantAgent.Request],
       evcs: UUID,
   )(implicit
-      serviceStateData: ExtEvStateData
+      serviceStateData: ExtEvStateData,
+      ctx: ActorContext[EvMessage],
   ): ExtEvStateData = {
-    log.debug(
+    ctx.log.debug(
       "Received ev movement service registration from {} for [Evcs:{}]",
       agentToBeRegistered.path.name,
       evcs,
@@ -157,7 +180,7 @@ class ExtEvDataService(override val scheduler: ActorRef)
         )
       case Some(_) =>
         // actor is already registered, do nothing
-        log.warning(
+        ctx.log.warn(
           "Sending actor {} is already registered",
           agentToBeRegistered,
         )
@@ -173,39 +196,47 @@ class ExtEvDataService(override val scheduler: ActorRef)
     *   the current state data of this service
     * @return
     *   the service stata data that should be used in the next state (normally
-    *   with updated values) together with the completion message that is sent
+    *   with updated values) together with the <completion> message that is sent
     *   in response to the trigger that was sent to start this announcement
     */
   override protected def announceInformation(
       tick: Long
-  )(implicit serviceStateData: ExtEvStateData, ctx: ActorContext): (
-      ExtEvStateData,
+  )(implicit
+      serviceStateData: S,
+      ctx: ActorContext[EvMessage],
+  ): (
+      S,
       Option[Long],
   ) = {
     def asScala[E]
         : java.util.Map[UUID, java.util.List[E]] => Map[UUID, Seq[E]] = map =>
       map.asScala.view.mapValues(_.asScala.toSeq).toMap
 
-    serviceStateData.extEvMessage.getOrElse(
-      throw ServiceException(
-        "ExtEvDataService was triggered without ExtEvMessage available"
-      )
-    ) match {
-      case _: RequestCurrentPrices =>
-        requestCurrentPrices()
-      case _: RequestEvcsFreeLots =>
-        requestFreeLots(tick)
-      case departingEvsRequest: RequestDepartingEvs =>
-        requestDepartingEvs(tick, asScala(departingEvsRequest.departures))
-      case arrivingEvsProvision: ProvideArrivingEvs =>
-        handleArrivingEvs(
-          tick,
-          asScala(arrivingEvsProvision.arrivals),
-          arrivingEvsProvision.maybeNextTick.toScala.map(Long2long),
-        )(
-          serviceStateData
+    implicit val log: Logger = ctx.log
+
+    serviceStateData.extEvMessage
+      .map {
+        case _: RequestCurrentPrices =>
+          requestCurrentPrices()
+        case _: RequestEvcsFreeLots =>
+          requestFreeLots(tick)
+        case departingEvsRequest: RequestDepartingEvs =>
+          requestDepartingEvs(tick, asScala(departingEvsRequest.departures))
+        case arrivingEvsProvision: ProvideArrivingEvs =>
+          handleArrivingEvs(
+            tick,
+            asScala(arrivingEvsProvision.arrivals),
+            arrivingEvsProvision.maybeNextTick.toScala.map(Long2long),
+          )(
+            serviceStateData,
+            ctx,
+          )
+      }
+      .getOrElse(
+        throw ServiceException(
+          "ExtEvDataService was triggered without ExtEvMessage available"
         )
-    }
+      )
   }
 
   private def requestCurrentPrices()(implicit
@@ -257,7 +288,8 @@ class ExtEvDataService(override val scheduler: ActorRef)
       tick: Long,
       requestedDepartingEvs: Map[UUID, Seq[UUID]],
   )(implicit
-      serviceStateData: ExtEvStateData
+      serviceStateData: ExtEvStateData,
+      log: Logger,
   ): (ExtEvStateData, Option[Long]) = {
 
     val departingEvResponses =
@@ -269,7 +301,7 @@ class ExtEvDataService(override val scheduler: ActorRef)
             Some(evcs)
 
           case None =>
-            log.warning(
+            log.warn(
               "A corresponding actor ref for UUID {} could not be found",
               evcs,
             )
@@ -297,12 +329,11 @@ class ExtEvDataService(override val scheduler: ActorRef)
       allArrivingEvs: Map[UUID, Seq[EvModel]],
       maybeNextTick: Option[Long],
   )(implicit
-      serviceStateData: ExtEvStateData
+      serviceStateData: ExtEvStateData,
+      ctx: ActorContext[EvMessage],
   ): (ExtEvStateData, Option[Long]) = {
 
     if (tick == INIT_SIM_TICK) {
-      // During initialization, an empty ProvideArrivingEvs message
-      // is sent, which includes the first relevant tick
 
       val nextTick = maybeNextTick.getOrElse(
         throw new CriticalFailureException(
@@ -312,7 +343,7 @@ class ExtEvDataService(override val scheduler: ActorRef)
 
       serviceStateData.uuidToActorRef.foreach { case (_, actor) =>
         actor ! RegistrationSuccessfulMessage(
-          self,
+          ctx.self.toClassic,
           nextTick,
         )
       }
@@ -324,7 +355,7 @@ class ExtEvDataService(override val scheduler: ActorRef)
 
         actor ! DataProvision(
           tick,
-          self,
+          ctx.self.toClassic,
           ArrivingEvs(evs.map(EvModelWrapper.apply)),
           maybeNextTick,
         )
@@ -342,7 +373,7 @@ class ExtEvDataService(override val scheduler: ActorRef)
 
   override protected def handleDataMessage(
       extMsg: DataMessageFromExt
-  )(implicit serviceStateData: ExtEvStateData): ExtEvStateData =
+  )(implicit serviceStateData: S): S =
     extMsg match {
       case extEvMessage: EvDataMessageFromExt =>
         serviceStateData.copy(