+++ src/main/scala/edu/ie3/simona/sim/setup/ExtSimSetupData.scala
@@ -6,14 +6,18 @@
 
 package edu.ie3.simona.sim.setup
 
+import edu.ie3.simona.ontology.messages.services.{EvMessage, ServiceMessage}
 import org.apache.pekko.actor.{ActorRef => ClassicRef}
 import edu.ie3.simona.service.ev.ExtEvDataService
+import org.apache.pekko.actor.typed.ActorRef
 
 final case class ExtSimSetupData(
     extSimAdapters: Iterable[ClassicRef],
-    extDataServices: Map[Class[_], ClassicRef],
+    extDataServices: Map[Class[_], ActorRef[_ >: ServiceMessage]],
 ) {
 
-  def evDataService: Option[ClassicRef] =
-    extDataServices.get(classOf[ExtEvDataService])
+  def evDataService: Option[ActorRef[EvMessage]] =
+    extDataServices
+      .get(classOf[ExtEvDataService.type])
+      .map { case service: ActorRef[EvMessage] => service }
 }