+++ src/main/scala/edu/ie3/simona/sim/setup/SimonaStandaloneSetup.scala
@@ -25,6 +25,7 @@ import edu.ie3.simona.exceptions.agent.GridAgentInitializationException
 import edu.ie3.simona.io.grid.GridProvider
 import edu.ie3.simona.ontology.messages.SchedulerMessage
 import edu.ie3.simona.ontology.messages.SchedulerMessage.ScheduleActivation
+import edu.ie3.simona.ontology.messages.services.ServiceMessage
 import edu.ie3.simona.scheduler.core.Core.CoreFactory
 import edu.ie3.simona.scheduler.core.RegularSchedulerCore
 import edu.ie3.simona.scheduler.{ScheduleLock, Scheduler, TimeAdvancer}
@@ -230,17 +231,19 @@ class SimonaStandaloneSetup(
           extLink.setup(extSimAdapterData)
           val extSim = extLink.getExtSimulation
 
-          val extDataInit
-              : Iterable[(Class[_ <: SimonaService[_]], ClassicRef)] =
+          val extDataInit: Iterable[(Class[_], ActorRef[_ >: ServiceMessage])] =
             extSim.getDataConnections.asScala.zipWithIndex.map {
               case (evConnection: ExtEvDataConnection, dIndex) =>
-                val extEvDataService = context.toClassic.simonaActorOf(
-                  ExtEvDataService.props(scheduler.toClassic),
-                  s"$index-$dIndex",
+                val extEvDataService = context.spawn(
+                  ExtEvDataService(scheduler),
+                  s"ExtEvDataService-$index-$dIndex",
                 )
-                evConnection.setActorRefs(extEvDataService, extSimAdapter)
+                evConnection.setActorRefs(
+                  extEvDataService.toClassic,
+                  extSimAdapter,
+                )
 
-                extEvDataService ! SimonaService.Create(
+                extEvDataService ! ServiceMessage.Create(
                   InitExtEvData(evConnection),
                   ScheduleLock.singleKey(
                     context,
@@ -249,7 +252,7 @@ class SimonaStandaloneSetup(
                   ),
                 )
 
-                (classOf[ExtEvDataService], extEvDataService)
+                (classOf[ExtEvDataService.type], extEvDataService)
             }
 
           // starting external simulation